# rest-server/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Paperless Rest Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/documents" : {
      "get" : {
        "tags" : [ "Documents" ],
        "operationId" : "GetDocuments",
        "parameters" : [ {
          "name" : "Page",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "page_size",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "query",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "ordering",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "tags__id__all",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int32"
            }
          }
        }, {
          "name" : "document_type__id",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "storage_path__id__in",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "correspondent__id",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "truncate_content",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_2"
                },
                "examples" : {
                  "default" : {
                    "value" : "{\"count\":1,\"next\":null,\"previous\":null,\"all\":[1],\"results\":[{\"id\":1,\"correspondent\":1,\"document_type\":1,\"storage_path\":1,\"title\":\"test\",\"content\":\"test\",\"tags\":[1],\"created\":\"2023-09-08T00:00:00Z\",\"created_date\":\"2023-09-08\",\"modified\":\"2023-09-08T12:15:24.652098Z\",\"added\":\"2023-09-08T12:05:28.338799Z\",\"archive_serial_number\":1,\"original_file_name\":\"test.pdf\",\"archived_file_name\":\"2023-09-08 edtest test.pdf\",\"owner\":3,\"user_can_change\":true,\"notes\":[{\"id\":1,\"note\":\"pdf in testdir\",\"created\":\"2023-09-08T12:11:05.848131Z\",\"document\":1,\"user\":3},{\"id\":2,\"note\":\"pdf 2 in test\",\"created\":\"2023-09-08T12:15:12.328015Z\",\"document\":1,\"user\":3}]}]}"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/documents/bulk_edit" : {
      "post" : {
        "tags" : [ "Documents" ],
        "operationId" : "BulkEdit",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/documents_bulk_edit_body"
              },
              "examples" : {
                "default" : {
                  "value" : "{\"documents\":[2],\"method\":\"redo_ocr\",\"parameters\":{}}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      }
    },
    "/api/documents/selection_data" : {
      "post" : {
        "tags" : [ "Documents" ],
        "operationId" : "SelectionData",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/documents_selection_data_body"
              },
              "examples" : {
                "default" : {
                  "value" : "{\"documents\":[]}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_5"
                },
                "examples" : {
                  "default" : {
                    "value" : "{\"selected_correspondents\":[{\"id\":1,\"document_count\":0}],\"selected_tags\":[{\"id\":1,\"document_count\":0}],\"selected_document_types\":[{\"id\":1,\"document_count\":0}],\"selected_storage_paths\":[{\"id\":1,\"document_count\":0}]}"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}/suggestions" : {
      "get" : {
        "tags" : [ "Documents" ],
        "operationId" : "GetDocumentSuggestions",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_6"
                },
                "examples" : {
                  "default" : {
                    "value" : "{\"correspondents\":[],\"tags\":[],\"document_types\":[],\"storage_paths\":[],\"dates\":[]}"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}/thumb" : {
      "get" : {
        "tags" : [ "Documents" ],
        "operationId" : "GetDocumentThumb",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/pdf" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary",
                  "x-content-type" : "application/pdf"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}/download" : {
      "get" : {
        "tags" : [ "Documents" ],
        "operationId" : "DownloadDocument",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "original",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/pdf" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary",
                  "x-content-type" : "application/pdf"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}/preview" : {
      "get" : {
        "tags" : [ "Documents" ],
        "operationId" : "GetDocumentPreview",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/pdf" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary",
                  "x-content-type" : "application/pdf"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}/metadata" : {
      "get" : {
        "tags" : [ "Documents" ],
        "operationId" : "GetDocumentMetadata",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_7"
                },
                "examples" : {
                  "default" : {
                    "value" : "{\"original_checksum\":\"39b80972dc7f1e975d9d2e8b92ce4e8b\",\"original_size\":11868,\"original_mime_type\":\"application/pdf\",\"media_filename\":\"testdirp.pdf\",\"has_archive_version\":true,\"original_metadata\":[],\"archive_checksum\":\"3597d60204a4b200b97e728622f704ab\",\"archive_media_filename\":\"testdirp.pdf\",\"original_filename\":\"test.pdf\",\"lang\":\"et\",\"archive_size\":15561,\"archive_metadata\":[{\"namespace\":\"http://ns.adobe.com/pdf/1.3/\",\"prefix\":\"pdf\",\"key\":\"Producer\",\"value\":\"pikepdf 7.2.0\"},{\"namespace\":\"http://ns.adobe.com/xap/1.0/\",\"prefix\":\"xmp\",\"key\":\"ModifyDate\",\"value\":\"2023-09-08T12:04:59+00:00\"},{\"namespace\":\"http://ns.adobe.com/xap/1.0/\",\"prefix\":\"xmp\",\"key\":\"CreateDate\",\"value\":\"2023-09-08T11:46:26+00:00\"},{\"namespace\":\"http://ns.adobe.com/xap/1.0/\",\"prefix\":\"xmp\",\"key\":\"CreatorTool\",\"value\":\"ocrmypdf 14.3.0 / Tesseract OCR-PDF 5.3.0\"},{\"namespace\":\"http://ns.adobe.com/xap/1.0/mm/\",\"prefix\":\"xmpMM\",\"key\":\"DocumentID\",\"value\":\"uuid:7af613b2-865c-11f9-0000-925a0e2aa608\"},{\"namespace\":\"http://purl.org/dc/elements/1.1/\",\"prefix\":\"dc\",\"key\":\"format\",\"value\":\"application/pdf\"},{\"namespace\":\"http://purl.org/dc/elements/1.1/\",\"prefix\":\"dc\",\"key\":\"title\",\"value\":\"Ohne Titel\"},{\"namespace\":\"http://www.aiim.org/pdfa/ns/id/\",\"prefix\":\"pdfaid\",\"key\":\"part\",\"value\":\"2\"},{\"namespace\":\"http://www.aiim.org/pdfa/ns/id/\",\"prefix\":\"pdfaid\",\"key\":\"conformance\",\"value\":\"B\"},{\"namespace\":\"http://purl.org/dc/elements/1.1/\",\"prefix\":\"dc\",\"key\":\"creator\",\"value\":\"None\"},{\"namespace\":\"http://ns.adobe.com/xap/1.0/\",\"prefix\":\"xmp\",\"key\":\"MetadataDate\",\"value\":\"2023-09-08T12:04:59.067651+00:00\"}]}"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "inline_response_200_2" : {
        "required" : [ "all", "count", "next", "previous", "results" ],
        "type" : "object",
        "properties" : {
          "count" : {
            "type" : "integer"
          },
          "next" : {
            "type" : "integer"
          },
          "previous" : {
            "type" : "integer"
          },
          "all" : {
            "type" : "array",
            "items" : {
              "type" : "integer"
            }
          },
          "results" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_2_results"
            }
          }
        },
        "example" : {
          "next" : 6,
          "all" : [ 5, 5 ],
          "previous" : 1,
          "count" : 0,
          "results" : [ {
            "owner" : 4,
            "user_can_change" : true,
            "archive_serial_number" : 2,
            "notes" : [ {
              "note" : "note",
              "created" : "created",
              "document" : 1,
              "id" : 7,
              "user" : 1
            }, {
              "note" : "note",
              "created" : "created",
              "document" : 1,
              "id" : 7,
              "user" : 1
            } ],
            "added" : "added",
            "created" : "created",
            "title" : "title",
            "content" : "content",
            "tags" : [ 3, 3 ],
            "storage_path" : 9,
            "archived_file_name" : "archived_file_name",
            "modified" : "modified",
            "correspondent" : 2,
            "original_file_name" : "original_file_name",
            "id" : 5,
            "created_date" : "created_date",
            "document_type" : 7
          }, {
            "owner" : 4,
            "user_can_change" : true,
            "archive_serial_number" : 2,
            "notes" : [ {
              "note" : "note",
              "created" : "created",
              "document" : 1,
              "id" : 7,
              "user" : 1
            }, {
              "note" : "note",
              "created" : "created",
              "document" : 1,
              "id" : 7,
              "user" : 1
            } ],
            "added" : "added",
            "created" : "created",
            "title" : "title",
            "content" : "content",
            "tags" : [ 3, 3 ],
            "storage_path" : 9,
            "archived_file_name" : "archived_file_name",
            "modified" : "modified",
            "correspondent" : 2,
            "original_file_name" : "original_file_name",
            "id" : 5,
            "created_date" : "created_date",
            "document_type" : 7
          } ]
        }
      },
...
      "documents_bulk_edit_body" : {
        "required" : [ "documents", "method", "parameters" ],
        "type" : "object",
        "properties" : {
          "documents" : {
            "type" : "array",
            "items" : {
              "type" : "integer"
            }
          },
          "method" : {
            "type" : "string"
          },
          "parameters" : {
            "type" : "object"
          }
        }
      },
      "documents_selection_data_body" : {
        "required" : [ "documents" ],
        "type" : "object",
        "properties" : {
          "documents" : {
            "type" : "array",
            "items" : {
              "type" : "integer"
            }
          }
        }
      },
      "inline_response_200_5" : {
        "required" : [ "selected_correspondents", "selected_document_types", "selected_storage_paths", "selected_tags" ],
        "type" : "object",
        "properties" : {
          "selected_correspondents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_5_selected_correspondents"
            }
          },
          "selected_tags" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_5_selected_correspondents"
            }
          },
          "selected_document_types" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_5_selected_correspondents"
            }
          },
          "selected_storage_paths" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_5_selected_correspondents"
            }
          }
        },
        "example" : {
          "selected_storage_paths" : [ null, null ],
          "selected_document_types" : [ null, null ],
          "selected_correspondents" : [ {
            "document_count" : 6,
            "id" : 0
          }, {
            "document_count" : 6,
            "id" : 0
          } ],
          "selected_tags" : [ null, null ]
        }
      },
      "inline_response_200_6" : {
        "required" : [ "correspondents", "dates", "document_types", "storage_paths", "tags" ],
        "type" : "object",
        "properties" : {
          "correspondents" : {
            "type" : "array",
            "items" : { }
          },
          "tags" : {
            "type" : "array",
            "items" : { }
          },
          "document_types" : {
            "type" : "array",
            "items" : { }
          },
          "storage_paths" : {
            "type" : "array",
            "items" : { }
          },
          "dates" : {
            "type" : "array",
            "items" : { }
          }
        },
        "example" : {
          "storage_paths" : [ "", "" ],
          "document_types" : [ "", "" ],
          "dates" : [ "", "" ],
          "correspondents" : [ "", "" ],
          "tags" : [ "", "" ]
        }
      },
...
      "inline_response_200_5_selected_correspondents" : {
        "required" : [ "document_count", "id" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "document_count" : {
            "type" : "integer"
          }
        },
        "example" : {
          "document_count" : 6,
          "id" : 0
        }
      },
...
    }
  }
}